requirements.txt
python-dotenv==1.0.0
streamlit==1.31.0
google-generativeai
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
PyPDF2==3.0.1